Specify timeout when using IAM-based credentials
We often get a timeout when using the EC2 metadata service to get AWS credentials from an IAM role, because the default timeout currently is 1s. This change adds a way to explictly tell KinesisWritable that we're using an IAM role to get credentials and also to specify a different timeout value.